Redo the top-level system code
