+++ .github/workflows/build.yml
@@ -79,7 +79,7 @@ jobs:
           sudo apt-get update -y
           sudo apt-get install -y libstdc++6
       - name: Run Kinoko
-        run: cd out && chmod u+x ./kinoko && ./kinoko -s testCases.bin
+        run: cd out && chmod u+x ./kinoko && ./kinoko -m test -s testCases.bin
       - name: Upload output
         uses: actions/upload-artifact@v4
         with:
