+++ source/host/KSystem.cc
@@ -1,0 +1,3 @@
+#include "KSystem.hh"
+
+KSystem *KSystem::s_instance = nullptr;
